+++ romancal/stpipe/core.py
@@ -2,6 +2,7 @@
 Roman Calibration Pipeline base class
 """
 
+import importlib.metadata
 import logging
 import time
 
@@ -16,6 +17,10 @@ _LOG_FORMATTER = logging.Formatter(
     datefmt="%Y-%m-%dT%H:%M:%S",
 )
 _LOG_FORMATTER.converter = time.gmtime
+
+
+log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
 
 
 class RomanStep(Step):
@@ -51,6 +56,8 @@ class RomanStep(Step):
             is the reftype code, the second element is the filename.
         """
 
+        model.meta.calibration_software_version = importlib.metadata.version("romancal")
+
         if isinstance(model, ImageModel):
             for log_record in self.log_records:
                 model.cal_logs.append(_LOG_FORMATTER.format(log_record))
@@ -60,14 +67,17 @@ class RomanStep(Step):
                 if hasattr(model.meta.ref_file, ref_name):
                     setattr(model.meta.ref_file, ref_name, ref_file)
                     # getattr(model.meta.ref_file, ref_name).name = ref_file
-
-        # this will only run if 'parent' is none, which happens when an individual
-        # step is being run or if self is a RomanPipeline and not a RomanStep.
-        if self.parent is None:
             model.meta.ref_file.crds.sw_version = crds_client.get_svn_version()
             model.meta.ref_file.crds.context_used = crds_client.get_context_used(
                 model.crds_observatory
             )
+
+            # this will only run if 'parent' is none, which happens when an individual
+            # step is being run or if self is a RomanPipeline and not a RomanStep.
+            if self.parent is None:
+                log.info(
+                    f"Results used CRDS context: {model.meta.ref_file.crds.context_used}"
+                )
 
     def record_step_status(self, model, step_name, success=True):
         """