+++ romancal/stpipe/tests/test_core.py
@@ -5,6 +5,8 @@ from roman_datamodels.datamodels import FlatRefModel, ImageModel
 from roman_datamodels.maker_utils import mk_level2_image
 from stpipe import crds_client
 
+import romancal
+from romancal.flatfield import FlatFieldStep
 from romancal.stpipe import RomanPipeline, RomanStep
 
 
@@ -82,17 +84,25 @@ def test_log_messages(tmp_path):
 def test_crds_meta():
     """Test that context and software versions are set"""
 
-    class ReflectStep(RomanStep):
-        def process(self, input):
-            return input
-
     im = ImageModel(mk_level2_image(shape=(20, 20)))
-    im.meta.ref_file.crds.sw_version = "junkversion"
-    im.meta.ref_file.crds.context_used = "junkcontext"
-
-    result = ReflectStep.call(im)
+    result = FlatFieldStep.call(im)
 
     assert result.meta.ref_file.crds.sw_version == crds_client.get_svn_version()
     assert result.meta.ref_file.crds.context_used == crds_client.get_context_used(
         result.crds_observatory
     )
+
+
+def test_calibration_software_version():
+    """Test that calibration_software_version is updated when a step is run"""
+
+    class NullStep(RomanStep):
+        def process(self, input):
+            return input
+
+    im = ImageModel(mk_level2_image(shape=(20, 20)))
+    im.meta.calibration_software_version = "junkversion"
+
+    result = NullStep.call(im)
+
+    assert result.meta.calibration_software_version == romancal.__version__